# Bugfix array equality