Elementwise comparison is no longer permitted when NumPy shapes cannot be broadcast (I need to find an official reference for this).

Much of Iris is already protected behind various `array_equal` calls, this PR mops up the remaining ones.

Targetting `v3.12.x` as this is required as an urgent bugfix for some users of Iris 3.12. Can propagate afterwards as necessary.