@HGWright 2025-08-22
Benchmark Report Action Failed so here are the results.



# :stopwatch: Performance Benchmark Report: 00e8b66a

<details>
<summary>Performance shifts</summary>

```

```

</details>

<details>
<summary>Full benchmark results</summary>

```

Benchmarks that have stayed the same:

| Change   | Before [ffaaab9f] <v3.12.2>   | After [00e8b66a]    |   Ratio | Benchmark (Parameter)                                                                       |
|----------|-------------------------------|---------------------|---------|---------------------------------------------------------------------------------------------|
|          | 21.1±0.2ms                    | 20.9±0.2ms          |    0.99 | aggregate_collapse.Aggregation.time_aggregated_by_COUNT(False)                              |
|          | 53.2±0.6ms                    | 52.7±0.4ms          |    0.99 | aggregate_collapse.Aggregation.time_aggregated_by_COUNT(True)                               |
|          | 35.3±0.3ms                    | 35.0±0.3ms          |    0.99 | aggregate_collapse.Aggregation.time_aggregated_by_FAST_PERCENTILE(False)                    |
|          | 167±2ms                       | 166±3ms             |    0.99 | aggregate_collapse.Aggregation.time_aggregated_by_FAST_PERCENTILE(True)                     |
|          | 23.3±0.2ms                    | 23.2±0.3ms          |    1    | aggregate_collapse.Aggregation.time_aggregated_by_GMEAN(False)                              |
|          | 31.5±0.3ms                    | 31.5±0.5ms          |    1    | aggregate_collapse.Aggregation.time_aggregated_by_GMEAN(True)                               |
|          | 23.6±0.3ms                    | 23.2±0.4ms          |    0.99 | aggregate_collapse.Aggregation.time_aggregated_by_HMEAN(False)                              |
|          | 31.9±0.1ms                    | 31.5±0.3ms          |    0.99 | aggregate_collapse.Aggregation.time_aggregated_by_HMEAN(True)                               |
|          | 21.2±0.2ms                    | 20.7±0.2ms          |    0.98 | aggregate_collapse.Aggregation.time_aggregated_by_MAX(False)                                |
|          | 45.9±0.5ms                    | 45.5±0.7ms          |    0.99 | aggregate_collapse.Aggregation.time_aggregated_by_MAX(True)                                 |
|          | 127±0.7ms                     | 126±1ms             |    1    | aggregate_collapse.Aggregation.time_aggregated_by_MAX_RUN(False)                            |
|          | 129±1ms                       | 130±0.7ms           |    1.01 | aggregate_collapse.Aggregation.time_aggregated_by_MAX_RUN(True)                             |
|          | 22.1±0.2ms                    | 22.4±0.2ms          |    1.01 | aggregate_collapse.Aggregation.time_aggregated_by_MEAN(False)                               |
|          | 48.1±0.5ms                    | 48.9±0.6ms          |    1.02 | aggregate_collapse.Aggregation.time_aggregated_by_MEAN(True)                                |
|          | 23.5±0.4ms                    | 23.9±0.2ms          |    1.02 | aggregate_collapse.Aggregation.time_aggregated_by_MEDIAN(False)                             |
|          | 57.7±0.9ms                    | 57.7±0.7ms          |    1    | aggregate_collapse.Aggregation.time_aggregated_by_MEDIAN(True)                              |
|          | 20.9±0.2ms                    | 21.0±0.1ms          |    1.01 | aggregate_collapse.Aggregation.time_aggregated_by_MIN(False)                                |
|          | 45.5±0.7ms                    | 45.0±0.5ms          |    0.99 | aggregate_collapse.Aggregation.time_aggregated_by_MIN(True)                                 |
|          | 1.09±0.01s                    | 1.09±0s             |    1    | aggregate_collapse.Aggregation.time_aggregated_by_PEAK(False)                               |
|          | 1.07±0.01s                    | 1.07±0.01s          |    1.01 | aggregate_collapse.Aggregation.time_aggregated_by_PEAK(True)                                |
|          | 215±5ms                       | 217±2ms             |    1.01 | aggregate_collapse.Aggregation.time_aggregated_by_PERCENTILE(False)                         |
|          | 349±6ms                       | 353±9ms             |    1.01 | aggregate_collapse.Aggregation.time_aggregated_by_PERCENTILE(True)                          |
|          | 21.9±0.6ms                    | 22.3±0.1ms          |    1.02 | aggregate_collapse.Aggregation.time_aggregated_by_PROPORTION(False)                         |
|          | 30.1±0.4ms                    | 30.0±0.4ms          |    1    | aggregate_collapse.Aggregation.time_aggregated_by_PROPORTION(True)                          |
|          | 22.4±0.3ms                    | 22.2±0.3ms          |    0.99 | aggregate_collapse.Aggregation.time_aggregated_by_RMS(False)                                |
|          | 60.5±0.8ms                    | 60.4±0.6ms          |    1    | aggregate_collapse.Aggregation.time_aggregated_by_RMS(True)                                 |
|          | 22.9±0.5ms                    | 23.3±0.2ms          |    1.02 | aggregate_collapse.Aggregation.time_aggregated_by_STD_DEV(False)                            |
|          | 62.2±0.5ms                    | 62.8±0.8ms          |    1.01 | aggregate_collapse.Aggregation.time_aggregated_by_STD_DEV(True)                             |
|          | 23.1±0.4ms                    | 23.2±0.2ms          |    1    | aggregate_collapse.Aggregation.time_aggregated_by_VARIANCE(False)                           |
|          | 59.3±0.7ms                    | 58.9±0.8ms          |    0.99 | aggregate_collapse.Aggregation.time_aggregated_by_VARIANCE(True)                            |
|          | 7.83±0.06ms                   | 7.66±0.1ms          |    0.98 | aggregate_collapse.Aggregation.time_collapsed_by_COUNT(False)                               |
|          | 21.8±0.4ms                    | 22.0±0.1ms          |    1.01 | aggregate_collapse.Aggregation.time_collapsed_by_COUNT(True)                                |
|          | 19.3±0.3ms                    | 19.0±0.2ms          |    0.98 | aggregate_collapse.Aggregation.time_collapsed_by_FAST_PERCENTILE(False)                     |
|          | 119±1ms                       | 120±2ms             |    1.01 | aggregate_collapse.Aggregation.time_collapsed_by_FAST_PERCENTILE(True)                      |
|          | 8.13±0.1ms                    | 8.00±0.1ms          |    0.98 | aggregate_collapse.Aggregation.time_collapsed_by_GMEAN(False)                               |
|          | 20.2±0.2ms                    | 19.8±0.2ms          |    0.98 | aggregate_collapse.Aggregation.time_collapsed_by_GMEAN(True)                                |
|          | 7.99±0.1ms                    | 7.99±0.07ms         |    1    | aggregate_collapse.Aggregation.time_collapsed_by_HMEAN(False)                               |
|          | 19.9±0.3ms                    | 19.7±0.3ms          |    0.99 | aggregate_collapse.Aggregation.time_collapsed_by_HMEAN(True)                                |
|          | 7.51±0.1ms                    | 7.56±0.05ms         |    1.01 | aggregate_collapse.Aggregation.time_collapsed_by_MAX(False)                                 |
|          | 20.7±0.4ms                    | 20.3±0.2ms          |    0.98 | aggregate_collapse.Aggregation.time_collapsed_by_MAX(True)                                  |
|          | 23.8±0.2ms                    | 23.7±0.2ms          |    0.99 | aggregate_collapse.Aggregation.time_collapsed_by_MAX_RUN(False)                             |
|          | 34.1±0.3ms                    | 34.6±0.4ms          |    1.02 | aggregate_collapse.Aggregation.time_collapsed_by_MAX_RUN(True)                              |
|          | 7.79±0.09ms                   | 7.76±0.1ms          |    1    | aggregate_collapse.Aggregation.time_collapsed_by_MEAN(False)                                |
|          | 21.0±0.1ms                    | 20.9±0.2ms          |    0.99 | aggregate_collapse.Aggregation.time_collapsed_by_MEAN(True)                                 |
|          | 9.26±0.2ms                    | 9.33±0.07ms         |    1.01 | aggregate_collapse.Aggregation.time_collapsed_by_MEDIAN(False)                              |
|          | 23.5±0.4ms                    | 23.3±0.3ms          |    0.99 | aggregate_collapse.Aggregation.time_collapsed_by_MEDIAN(True)                               |
|          | 7.54±0.08ms                   | 7.75±0.1ms          |    1.03 | aggregate_collapse.Aggregation.time_collapsed_by_MIN(False)                                 |
|          | 20.5±0.2ms                    | 20.6±0.4ms          |    1    | aggregate_collapse.Aggregation.time_collapsed_by_MIN(True)                                  |
|          | 529±3ms                       | 529±4ms             |    1    | aggregate_collapse.Aggregation.time_collapsed_by_PEAK(False)                                |
|          | 541±4ms                       | 536±2ms             |    0.99 | aggregate_collapse.Aggregation.time_collapsed_by_PEAK(True)                                 |
|          | 44.8±0.8ms                    | 45.1±0.3ms          |    1.01 | aggregate_collapse.Aggregation.time_collapsed_by_PERCENTILE(False)                          |
|          | 132±3ms                       | 129±2ms             |    0.98 | aggregate_collapse.Aggregation.time_collapsed_by_PERCENTILE(True)                           |
|          | 7.86±0.1ms                    | 7.87±0.1ms          |    1    | aggregate_collapse.Aggregation.time_collapsed_by_PROPORTION(False)                          |
|          | 19.8±0.4ms                    | 20.0±0.3ms          |    1.01 | aggregate_collapse.Aggregation.time_collapsed_by_PROPORTION(True)                           |
|          | 7.91±0.07ms                   | 7.84±0.1ms          |    0.99 | aggregate_collapse.Aggregation.time_collapsed_by_RMS(False)                                 |
|          | 22.5±0.3ms                    | 22.6±0.3ms          |    1    | aggregate_collapse.Aggregation.time_collapsed_by_RMS(True)                                  |
|          | 8.03±0.1ms                    | 8.06±0.1ms          |    1    | aggregate_collapse.Aggregation.time_collapsed_by_STD_DEV(False)                             |
|          | 22.3±0.3ms                    | 22.4±0.4ms          |    1.01 | aggregate_collapse.Aggregation.time_collapsed_by_STD_DEV(True)                              |
|          | 8.17±0.1ms                    | 8.08±0.09ms         |    0.99 | aggregate_collapse.Aggregation.time_collapsed_by_VARIANCE(False)                            |
|          | 22.0±0.2ms                    | 21.6±0.4ms          |    0.98 | aggregate_collapse.Aggregation.time_collapsed_by_VARIANCE(True)                             |
|          | 22.7±0.1ms                    | 22.6±0.2ms          |    1    | aggregate_collapse.WeightedAggregation.time_w_aggregated_by_MEAN(False)                     |
|          | 83.1±0.6ms                    | 82.5±0.7ms          |    0.99 | aggregate_collapse.WeightedAggregation.time_w_aggregated_by_MEAN(True)                      |
|          | 22.5±0.3ms                    | 22.7±0.2ms          |    1.01 | aggregate_collapse.WeightedAggregation.time_w_aggregated_by_RMS(False)                      |
|          | 95.4±1ms                      | 96.0±0.4ms          |    1.01 | aggregate_collapse.WeightedAggregation.time_w_aggregated_by_RMS(True)                       |
|          | 21.1±0.2ms                    | 21.6±0.2ms          |    1.02 | aggregate_collapse.WeightedAggregation.time_w_aggregated_by_SUM(False)                      |
|          | 56.1±0.7ms                    | 56.4±0.5ms          |    1.01 | aggregate_collapse.WeightedAggregation.time_w_aggregated_by_SUM(True)                       |
|          | 8.28±0.1ms                    | 8.23±0.06ms         |    0.99 | aggregate_collapse.WeightedAggregation.time_w_collapsed_by_MEAN(False)                      |
|          | 26.7±0.4ms                    | 26.9±0.5ms          |    1.01 | aggregate_collapse.WeightedAggregation.time_w_collapsed_by_MEAN(True)                       |
|          | 8.04±0.08ms                   | 8.11±0.06ms         |    1.01 | aggregate_collapse.WeightedAggregation.time_w_collapsed_by_RMS(False)                       |
|          | 28.3±0.5ms                    | 28.3±0.7ms          |    1    | aggregate_collapse.WeightedAggregation.time_w_collapsed_by_RMS(True)                        |
|          | 7.89±0.08ms                   | 7.82±0.09ms         |    0.99 | aggregate_collapse.WeightedAggregation.time_w_collapsed_by_SUM(False)                       |
|          | 23.2±0.6ms                    | 23.3±0.7ms          |    1    | aggregate_collapse.WeightedAggregation.time_w_collapsed_by_SUM(True)                        |
|          | 217±1ms                       | 217±2ms             |    1    | aggregate_collapse.WeightedAggregation.time_w_collapsed_by_WPERCENTILE(False)               |
|          | 268±3ms                       | 266±6ms             |    0.99 | aggregate_collapse.WeightedAggregation.time_w_collapsed_by_WPERCENTILE(True)                |
|          | 1.09±0.02ms                   | 1.10±0.01ms         |    1.01 | cube.CubeCreation.time_create(False, 'construct')                                           |
|          | 385±5μs                       | 389±6μs             |    1.01 | cube.CubeCreation.time_create(False, 'instantiate')                                         |
|          | 929±9μs                       | 930±10μs            |    1    | cube.CubeCreation.time_create(True, 'construct')                                            |
|          | 558±6μs                       | 562±10μs            |    1.01 | cube.CubeCreation.time_create(True, 'instantiate')                                          |
|          | 84.5±1ms                      | 85.3±1ms            |    1.01 | cube.CubeEquality.time_equality(False, False, 'all_equal')                                  |
|          | 26.2±0.9ms                    | 25.6±0.5ms          |    0.98 | cube.CubeEquality.time_equality(False, False, 'coord_inequality')                           |
|          | 98.8±1ms                      | 98.9±1ms            |    1    | cube.CubeEquality.time_equality(False, False, 'data_inequality')                            |
|          | 16.3±0.3μs                    | 16.8±0.2μs          |    1.03 | cube.CubeEquality.time_equality(False, False, 'metadata_inequality')                        |
|          | 85.6±1ms                      | 84.9±1ms            |    0.99 | cube.CubeEquality.time_equality(False, True, 'all_equal')                                   |
|          | 25.8±0.8ms                    | 25.9±0.3ms          |    1.01 | cube.CubeEquality.time_equality(False, True, 'coord_inequality')                            |
|          | 101±2ms                       | 99.8±2ms            |    0.99 | cube.CubeEquality.time_equality(False, True, 'data_inequality')                             |
|          | 16.4±0.2μs                    | 16.4±0.1μs          |    1    | cube.CubeEquality.time_equality(False, True, 'metadata_inequality')                         |
|          | 171±2ms                       | 174±2ms             |    1.01 | cube.CubeEquality.time_equality(True, False, 'all_equal')                                   |
|          | 67.2±1ms                      | 67.9±1ms            |    1.01 | cube.CubeEquality.time_equality(True, False, 'coord_inequality')                            |
|          | 197±4ms                       | 199±1ms             |    1.01 | cube.CubeEquality.time_equality(True, False, 'data_inequality')                             |
|          | 51.2±0.2μs                    | 51.2±0.8μs          |    1    | cube.CubeEquality.time_equality(True, False, 'metadata_inequality')                         |
|          | 242±6ms                       | 238±2ms             |    0.99 | cube.CubeEquality.time_equality(True, True, 'all_equal')                                    |
|          | 135±2ms                       | 138±3ms             |    1.02 | cube.CubeEquality.time_equality(True, True, 'coord_inequality')                             |
|          | 267±4ms                       | 269±6ms             |    1.01 | cube.CubeEquality.time_equality(True, True, 'data_inequality')                              |
|          | 53.1±0.5μs                    | 53.4±0.9μs          |    1.01 | cube.CubeEquality.time_equality(True, True, 'metadata_inequality')                          |
|          | 805±4μs                       | 803±6μs             |    1    | import_iris.Iris.time__concatenate                                                          |
|          | 189±3μs                       | 187±8μs             |    0.99 | import_iris.Iris.time__constraints                                                          |
|          | 115±2μs                       | 115±0.8μs           |    1    | import_iris.Iris.time__data_manager                                                         |
|          | 92.4±2μs                      | 91.2±0.8μs          |    0.99 | import_iris.Iris.time__deprecation                                                          |
|          | 163±1μs                       | 164±3μs             |    1.01 | import_iris.Iris.time__lazy_data                                                            |
|          | 909±20μs                      | 908±3μs             |    1    | import_iris.Iris.time__merge                                                                |
|          | 74.1±0.3μs                    | 74.7±0.8μs          |    1.01 | import_iris.Iris.time__representation                                                       |
|          | 624±7μs                       | 613±10μs            |    0.98 | import_iris.Iris.time_analysis                                                              |
|          | 139±1μs                       | 141±2μs             |    1.01 | import_iris.Iris.time_analysis__area_weighted                                               |
|          | 106±0.6μs                     | 108±0.9μs           |    1.01 | import_iris.Iris.time_analysis__grid_angles                                                 |
|          | 249±2μs                       | 246±2μs             |    0.99 | import_iris.Iris.time_analysis__interpolation                                               |
|          | 193±3μs                       | 192±3μs             |    0.99 | import_iris.Iris.time_analysis__regrid                                                      |
|          | 112±2μs                       | 112±2μs             |    1    | import_iris.Iris.time_analysis__scipy_interpolate                                           |
|          | 137±0.6μs                     | 138±1μs             |    1    | import_iris.Iris.time_analysis_calculus                                                     |
|          | 323±3μs                       | 325±2μs             |    1.01 | import_iris.Iris.time_analysis_cartography                                                  |
|          | 91.1±0.4μs                    | 93.0±2μs            |    1.02 | import_iris.Iris.time_analysis_geomerty                                                     |
|          | 212±1μs                       | 215±3μs             |    1.01 | import_iris.Iris.time_analysis_maths                                                        |
|          | 94.5±0.9μs                    | 95.9±0.8μs          |    1.02 | import_iris.Iris.time_analysis_stats                                                        |
…